add refresh exchange automatic

+++ resources/views/web/store.blade.php
@@ -128,7 +128,7 @@
                                             @if($store->total_views<=1000)
                                                 {{$store->total_views}}
                                                 @else
-                                                {{$store->total_views/1000}} K
+                                                {{(int)( $store->total_views/1000)}} K
                                                 @endif
                                         </p>
                                         <p class="sub-title">مشاهدات</p>
